requirements.txt
smtplib
email
schedule
python-dotenv
